refactor: isolate SQL connection workspace
