--- src/types/savedQuery.ts
@@ -1,9 +1,0 @@
-export interface SavedQuery {
-  id: number;
-  connection_uuid: string;
-  name: string;
-  query: string;
-  created_at: string;
-  updated_at: string;
-}
-
